refactor(Package): rename package domain to entity
Changed the entity as a name to make it look more appropriate compared to the domain and to be consistent with the order/entity's package name.
