+++ src/main/java/personal/delivery/cart/service/CartService.java
@@ -3,7 +3,7 @@ package personal.delivery.cart.service;
 import personal.delivery.cart.dto.CartMenuDto;
 import personal.delivery.cart.dto.CartMenuResponseDto;
 import personal.delivery.cart.entity.Cart;
-import personal.delivery.member.domain.Member;
+import personal.delivery.member.eneity.Member;
 
 public interface CartService {
