+++ src/main/java/personal/delivery/config/BeanConfiguration.java
@@ -4,7 +4,7 @@ import org.modelmapper.ModelMapper;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import personal.delivery.cart.entity.Cart;
-import personal.delivery.member.domain.Member;
+import personal.delivery.member.eneity.Member;
 import personal.delivery.menu.Menu;
 import personal.delivery.order.entity.Order;
