+++ src/main/java/personal/delivery/member/dto/MemberResponseDto.java
@@ -2,7 +2,7 @@ package personal.delivery.member.dto;
 
 import lombok.Data;
 import personal.delivery.constant.Role;
-import personal.delivery.member.domain.Address;
+import personal.delivery.member.eneity.Address;
 
 import java.time.LocalDateTime;
