+++ src/main/java/personal/delivery/member/eneity/Member.java
@@ -1,17 +1,17 @@
-package personal.delivery.member.domain;
+package personal.delivery.member.eneity;
 
 import jakarta.persistence.*;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import personal.delivery.constant.Role;
-import personal.delivery.member.domain.Address;
 
 import java.time.LocalDateTime;
 
 @Entity
 @Getter
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(name = "member")
 public class Member {
