+++ src/main/java/personal/delivery/member/repository/MemberRepository.java
@@ -1,7 +1,7 @@
 package personal.delivery.member.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import personal.delivery.member.domain.Member;
+import personal.delivery.member.eneity.Member;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
