+++ src/main/java/personal/delivery/order/entity/Order.java
@@ -2,11 +2,12 @@ package personal.delivery.order.entity;
 
 
 import jakarta.persistence.*;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import personal.delivery.constant.OrderStatus;
-import personal.delivery.member.domain.Member;
+import personal.delivery.member.eneity.Member;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -14,7 +15,7 @@ import java.util.List;
 
 @Entity
 @Getter
-@NoArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Table(name = "orders")
 public class Order {
